+++ egui/src/lib.rs
@@ -92,7 +92,10 @@ pub use {
     layout::*,
     math::*,
     memory::Memory,
-    paint::{color, PaintCmd, PaintJobs, Rgba, Srgba, Stroke, TextStyle, Texture, TextureId},
+    paint::{
+        color, FontDefinitions, FontFamily, PaintCmd, PaintJobs, Rgba, Srgba, Stroke, TextStyle,
+        Texture, TextureId,
+    },
     painter::Painter,
     style::Style,
     types::*,