+++ egui/src/paint/mod.rs
@@ -13,7 +13,7 @@ mod texture_atlas;
 pub use {
     color::{Rgba, Srgba},
     command::{PaintCmd, Stroke},
-    fonts::{FontDefinitions, Fonts, TextStyle},
+    fonts::{FontDefinitions, FontFamily, Fonts, TextStyle},
     stats::PaintStats,
     tessellator::{PaintJob, PaintJobs, PaintOptions, TextureId, Triangles, Vertex, WHITE_UV},
     texture_atlas::Texture,